+++ src/objects/hittable.rs
@@ -5,9 +5,9 @@ use crate::types::ray::Ray;
 pub struct HitRecord {
     pub p: Point,
     pub normal: Vec3,
-    pub t: f32
+    pub t: f64
 }
 
 pub trait Hittable {
-    fn hit(&self, r: &Ray, ray_tmin: f32, ray_tmax: f32) -> Option<HitRecord>; 
+    fn hit(&self, r: &Ray, ray_tmin: f64, ray_tmax: f64) -> Option<HitRecord>; 
 }
